Add Hopper TMA-pipelined forward ring kernel; lift BDIM_Y from launcher template
  attention_cuda_utils.cu / .cuh:
    * getPtxver() exposed via the .cuh (was static, now callable from
      the ring TUs). Returns attrs.ptxVersion directly. The previous
      `* 10` was wrong, since ptxVersion is already in major*10+minor
      form (compute_90 == 90). Caller-side comparisons against 90 now
      have the right scale.
    * Added ensure_dyn_shmem(const void*, size_t): one-shot
      cudaFuncSetAttribute(MaxDynamicSharedMemorySize) for kernels that
      need >48 KB of dynamic shared memory, with a function-local
      unordered_set cache so each kernel only pays the setup once.
      Required by the new TMA kernel.

  attention_cuda_fwd.cu:
    * In s2_attention_fwd_cuda, nchans_in/out: size_t -> int64_t to
      match the kernel argument types and avoid unsigned narrowing
      warning from nvcc.

  attention_cuda_bwd.cu:
    * Two `if (CHOUT_AS_IN)` -> `if constexpr(CHOUT_AS_IN)` in
      s2_attn_bwd_special_vec_k.

  attention_cuda_fwd_ring.cu:
    Renamed s2_attn_fwd_ring_step_special_vec_k -> ..._ldg_k (kept as
    the LDG path) and added a new *_tma_k variant for Hopper+:
      * double-buffered TMA prefetch of kx/vx neighbor vectors via
        cuda::device::memcpy_async_tx + cuda::device::barrier_expect_tx,
        with shared-memory cuda::barrier<thread_scope_block>.
      * Body wrapped in
          #if defined(__CUDA_ARCH__) && __CUDA_ARCH__ >= 900
            ...real implementation...
          #else
            __trap();
          #endif
        so the kernel still compiles in pre-Hopper passes for fat
        binaries; the trap is unreachable at runtime by construction.

    launch_spc_attn_ring_fwd:
      * Dropped the BDIM_Y template parameter; now derived as
        inside function.
      * Computes use_tma at runtime:
          const bool use_tma =
              std::is_same_v<FLOATV_T, float4> && (getPtxver() >= 90);
        VEC4 captures TMA's 16B alignment requirement; ptxVersion is
        the actual gate (cudaFuncGetAttributes resolves it against
        the loaded cubin/PTX for the current device, so it's correct
        across SASS-only, PTX-JIT, and multi-arch builds without a
        separate runtime CC check).

  attention_cuda_bwd_ring.cu:

    Parallel cleanup, no math change:
      * launch_spc_attn_ring_pass1_bwd / pass2_bwd: BDIM_Y dropped
        from the template parameter list, derived inside as above.
      * pass1_dispatch / pass2_dispatch: BDIM_Y dropped from every
        launcher callsite.
      * Removed unused NLOC_M1 in the special_vec kernels.

+++ torch_harmonics/attention/optimized/kernels_cuda/attention_cuda_fwd.cu
@@ -541,8 +541,8 @@ torch::Tensor s2_attention_fwd_cuda(at::Tensor kx,
     TORCH_CHECK(nlon_in % nlon_out == 0,
                 "nlon_in (", nlon_in, ") must be an integer multiple of nlon_out (", nlon_out, ")");
 
-    size_t nchans_in  = qy.size(1); // or kx.size(1)
-    size_t nchans_out = vx.size(1);
+    int64_t nchans_in  = qy.size(1); // or kx.size(1)
+    int64_t nchans_out = vx.size(1);
 
     const int batch_size = kx.size(0);
